requirements.txt
asgiref==3.10.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.3
Django==5.2.8
django-allauth==65.13.1
django-cors-headers==4.9.0
django-environ==0.12.0
django-otp==1.6.3
djangorestframework==3.16.1
gunicorn==23.0.0
idna==3.11
oauthlib==3.3.1
packaging==25.0
pillow==12.0.0
pycparser==2.23
pydotenv==0.0.7
PyJWT==2.10.1
python-decouple==3.8
qrcode==8.2
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.6.2
whitenoise==6.11.0
